ENH: Avoid template type specification for image file reader in Python wrapping
If no template parameter is specified when opening an image in Python, the image is read with itk.ImageIOFactory.CreateImageIO. The dimension, the component type and the pixel type are assessed from this first reader. The corresponding image type is created before instantiating a reader. The supported component and pixel types are based on the list available in ImageIOBase [1].
This paradigm is the same then the one that is widely used in C++ to open the image using the appropriate type [2].

[1] http://www.itk.org/Doxygen/html/classitk_1_1ImageIOBase.html
[2] https://github.com/Slicer/Slicer/blob/master/Modules/CLI/AddScalarVolumes/AddScalarVolumes.cxx#L107-L162

Change-Id: I07038689b50dbdfc0d81c5e08e05f5e8dd2cce10