BUG: lazyAttributes dictionary keys could be over-ridden
When an attribute belonged to multiple modules, only the last module
loaded was kept. To avoid this issue, the dictionary containing all
the attributes now saves a list of module for each attribute.
Additionally, only the main module containing an attribute is
loaded at first. The other modules are loaded only when required.

Change-Id: Ie072f46dac59c06749f89ddfc095063d5826c00b