ENH: Improving API of Graft() in itk::Image
Graft() function in the class Image requires an input of type
Image. Previously, if the given input of type DataObject was not
actually of type Image, an eception was thrown. The correct
implementation is to only allow a developer to pass an Image
as argument for Graft() to get an error at compile time instead
of getting an error at runtime.
Graft(const DataObject*) is now a protected function to allow
polymorphic manipulation of generic DataObjects via the base
class virtual method.

Change-Id: I91fd1cc809715b974b2649e5f155df3c48a69506