WIP: Addition of type long long and unsigned long long for images
ITK images can be templated over any type of pixel. However, the
image readers and writers need to support the type that is used
to be able to load and save files. Currently, the type of pixels
defined in itkImageIOBase.h that are supported does not include
'long long'. The types that are supported are: (u)char, (u)short,
(u)int, (u)long, float, and double. This does not garanty the support
of 64 bit images on certain platform as the size of long may vary.
Additionally, if one uses the type '(u)int64' or any of its derived
types, the writer will default on UNKNOWNCOMPONENTTYPE if the platform
defines '(u)int64' as '(unsigned) long long' (e.g. on Windows/Visual Studio
C++ 2013).

Change-Id: I01c3b0a4f01974f58545d5dc70a05f47d1adc3d5