COMP: DCMTK by-products not specified
When compiling modules that depended on DCMTK, DCMTK
was not build first. The by-products (libraries) that
were compiled in DCMTK were not declared in
ExternalProject_Add. Support for this argument is
only supported in CMake 3.2 and above. To fix this
issue, the ExternalProject module has been back-ported
and added to ITK.

Change-Id: I318262e097a2074a8054f18b15e310faeabb92f5
Inspired-by: Brad King <brad.king@kitware.com>