BUG: Replacing typeid() by a user-defined 'typeid' function
On MacOS, when compiling ITK with BUILD_SHARED_LIBS=ON and ITK_WRAP_PYTHON,
typeinfo of the ITK types needs to be visible outside of the library and must
be consistent across all translation units. In case of an inconsistency, the
linker will print a warning message such as:

ld: warning: direct access in itk::SimpleDataObjectDecorator<std::__1::vector\
<unsigned long, std::__1::allocator<unsigned long> > >::PrintSelf(\
std::__1::basic_ostream<char, std::__1::char_traits<char> >&, itk::Indent)\
const to global weak symbol typeinfo name for std::__1::vector<unsigned long,\
std::__1::allocator<unsigned long> > means the weak symbol cannot be overridden\
at runtime. This was likely caused by different translation units being\
compiled with different visibility settings.

typeid should not have to worry about the typeinfo being defined multiple times,
but there is no easy way to remove this warning without remove the usage of
typeid on MacOS when BUILD_SHARED_LIBS=ON for the types for which the visibility
is difficult to control (outside of ITK).

This defines two functions that are called to print the typeid:
-One for itkLightObjects which will handle most ITK objects.
-One for everything else, including vnl objects and std objects.

Suggested-by: Bradley Lowekamp <blowekamp@mail.nih.gov>

Change-Id: I7cc9e3b75c61f07e1ca737b260e7854639d52408