BUG: Alpha value created when converted image to RGBA is not always 1
When converting an image (scalar or RGB) to an RGBA image using the 'Convert()'
function available in itkConvertPixelBuffer.h, alpha values are filled
automatically. If the image components are 'double' or 'float', alpha must
be equal to 1. However, if the component type is some integer, alpha value
must be the largest possible value that the type accepts (e.g. for unsigned
char, alpha=255).

Change-Id: I929e8ff771c9d73a0fdde4e6582427ceacf717fb