ENH: Remove 'weak symbol' warning on MacOS
'weak symbol' warnings appear when compiling ITK on MacOS with
BUILD_SHARED_LIBS:BOOL=OFF and ITK_WRAP_PYTHON:BOOL=ON:

ld: warning: direct access in _wrap_itkMersenneTwisterRandomVariateGenerator\
_cast(_object*, _object*) to global weak symbol typeinfo for\
itk::Statistics::MersenneTwisterRandomVariateGenerator means the weak symbol\
cannot be overridden at runtime. This was likely caused by different\
translation units being compiled with different visibility settings.

This is due to export attributes being different some classes when they
are declared and forward-declared. New export macros were created to
be used when a class is forward-declared to match the attributes
in the actual declaration of the class.

The new export macros are:
${itk_module}_FORWARD_EXPORT
${itk_module}_FORWARD_HIDDEN
ITK_TEMPLATE_FORWARD_EXPORT

These macros are not empty only on Mac as it is the only plateform on which
the warnings arise, and adding attributes on the forward declaration
causes other warnings with certain versions of GCC (see commit
79095080f3e69f7014103045e039621fad5a2d83).

Change-Id: I5e4b45e4afac988ec68669b28237fa4819214e14