BUG: ICU_ARGS CMake variable should be declared CACHE
Non-cache variables do not persist across CMake runs. Since ICU_ARGS is only
defined if DCMTK_USE_ICU is not defined, it will be defined once, on the first
run, but on each subsequent CMake run, it will not be defined anymore. This
variable is used to configure ICU. Since it won't be defined, ICU will be
configured with its default CMake options, which means that it might try to
USE the system ICU, which will lead to linking errors when trying to use
DCMTK is external projects.
Also, remove FATAL_ERROR on MSVC as DCMTK and ICU should work.
Change-Id: I3f636f043deb38abde8529818061622ffcc6bde7