ENH: Update DCMTK build instructions for Windows
DMCTK can be build on Windows as part of the normal ITK build process. DCMTK
can depend on ICU, which on Windows will be downloaded precompiled. ICU has
been precompiled on a limited number of compilers, and it may be necessary
to precompile it for more compiler. Build instructions are available in
an external project linked in README.md .

Change-Id: Ieddc35dbd8a1a4dbc2476862628b14e61b903671