BUG: Synchronize factories across modules in Python
Factory registration is done in each Python module independenlty.
This lead to issues such as:
* Calling RegisterOneFactory() in Python was not working. The function would
run but the factory was not properly registered.
* It was not possible to register a remote IO module in Python since it
was not possible to register the module. Only modules that were selected at
compile time were registered.

A solution similar to the one used to solve the global time stamp issue in [1]
was used to solve this problem. It is however not possible to use the exact
same solution which uses the global time stamp from ITKCommon for all modules.

The solution implemented in this commit synchronizes all the factories (both
internal and registered factories) across
the different modules. This means that
if factories are registered in a module during static initialization, these
modules will still be registered after synchronization. Image factory
registration is performed in ITKIOImageBase. After synchronization, the image
factories are still registered. Using the solution implemented to solve the
global time stamp problem would have replaced the registered factories in
ITKIOImageBase by the ones registered in ITKCommon, which would lead to no
image factory registered.

[1] 67702504e840a60bc968e2e09804879d9c6c5801

Change-Id: I17e4b167b15d0634a6a121dfd0225bc1623333f0